tests/wpt/tests/client-hints/sandbox/iframe-csp-same-origin.https.html

<!DOCTYPE html>
<html>

<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/client-hints/resources/export.js"></script>
<script src="resources/util.js"></script>

<script>

sandbox_iframe_test("",
                    expect_iframe_hints,
                    "CSP sandboxed iframe with same-origin flag does send client hint headers");

</script>

</body>
</html>
