third_party/blink/web_tests/virtual/disable-user-agent-client-hint-feature/external/wpt/client-hints/sandbox/iframe-csp-same-origin.https-expected.txt

This is a testharness.js-based test.
FAIL CSP sandboxed iframe with same-origin flag does send client hint headers assert_equals: message from opened frame expected "PASS" but got "FAIL Sec-CH-UA True None"
Harness: the test ran to completion.
